more files

+++ WHA-Framework/WHA-Framework/Common/StepLibrary/IndividualEntityReport.cs
@@ -5,8 +5,8 @@ namespace WHA_Framework.Common.StepLibrary
     public class IndividualEntityReport
     {
         public int TransID { get; set; }
-        public int Amount { get; set; }
-        public int CurrentBal { get; set; }
+        public double Amount { get; set; }
+        public double CurrentBal { get; set; }
         public string AcName { get; set; }
         public string TransType { get; set; }
     }
